# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UserController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "timtimmerman@email.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "UnsafePassword123!"
                  },
                  "firstname": {
                    "type": "string",
                    "example": "Tim"
                  },
                  "surname": {
                    "type": "string",
                    "example": "Timmerman"
                  },
                  "role": {
                    "type": "string",
                    "example": "user"
                  },
                  "location": {
                    "type": "string",
                    "example": "Enschede, Nederland"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "firstname",
                  "surname"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "timtimmerman@email.com"
                  },
                  "firstname": {
                    "type": "string",
                    "example": "Tim"
                  },
                  "surname": {
                    "type": "string",
                    "example": "Timmerman"
                  },
                  "location": {
                    "type": "string",
                    "nullable": true,
                    "example": "Brussels, Belgium"
                  },
                  "latitude": {
                    "type": "number",
                    "nullable": true,
                    "example": 50.8503
                  },
                  "longitude": {
                    "type": "number",
                    "nullable": true,
                    "example": 4.3517
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}/avatar": {
      "post": {
        "operationId": "UserController_addAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "binary": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_removeAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Trade2",
    "description": "API Documentation for Trade2",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}